Fix some warning on 64-bit architectures.
git-svn-id: http://svn.ipd.uka.de/repos/ParaStation/trunk/mgmt@1080 34291a86-62e8-0310-8159-f060f8a2ee4e

+++ lib/psport4/psport4.c
@@ -7,7 +7,7 @@
 /**
  * name: Description
  *
- * $Id: psport4.c,v 1.4 2002/06/12 14:19:02 hauke Exp $
+ * $Id: psport4.c,v 1.5 2002/06/13 17:31:49 eicker Exp $
  *
  * @author
  *         Jens Hauke <hauke@par-tec.de>
@@ -996,19 +996,19 @@ void ConfigureConnection( int fd )
     errno = 0;
     val = env_so_sndbuf;
     ret = setsockopt( fd, SOL_SOCKET, SO_SNDBUF, &val, sizeof(val));
-    DPRINT( 2, "setsockopt( %d, SOL_SOCKET, SO_SNDBUF, [%d], %d ) = %d : %s",
-	    fd, val, sizeof(val), ret, strerror( errno ));
+    DPRINT( 2, "setsockopt( %d, SOL_SOCKET, SO_SNDBUF, [%d], %ld ) = %d : %s",
+	    fd, val, (long) sizeof(val), ret, strerror( errno ));
 
     errno = 0;
     val = env_so_rcvbuf;
     ret = setsockopt( fd, SOL_SOCKET, SO_RCVBUF, &val, sizeof(val));
-    DPRINT( 2, "setsockopt( %d, SOL_SOCKET, SO_RCVBUF, [%d], %d ) = %d : %s",
-	    fd, val, sizeof(val), ret, strerror( errno ));
+    DPRINT( 2, "setsockopt( %d, SOL_SOCKET, SO_RCVBUF, [%d], %ld ) = %d : %s",
+	    fd, val, (long) sizeof(val), ret, strerror( errno ));
     errno = 0;
     val = env_tcp_nodelay;
     ret = setsockopt( fd, IPPROTO_TCP, TCP_NODELAY, &val, sizeof(val));
-    DPRINT( 2, "setsockopt( %d, IPPROTO_TCP, TCP_NODELAY, [%d], %d ) = %d : %s",
-	    fd, val, sizeof(val), ret, strerror( errno ));
+    DPRINT( 2, "setsockopt( %d, IPPROTO_TCP, TCP_NODELAY, [%d], %ld ) = %d : %s",
+	    fd, val, (long) sizeof(val), ret, strerror( errno ));
 }
 #endif
